[Storage] Implement count all and add total field to the found entity

+++ NBlast.Storage.Core/Index/IStorageReader.fs
@@ -3,5 +3,5 @@
 type IStorageReader = interface
     abstract member SearchByField : string -> int option -> int option -> LogDocumentHits
     abstract member FindAll: int option -> int option -> LogDocumentHits
-    abstract member CountAll: unit -> bigint
+    abstract member CountAll: unit -> int
     end
